Move startup to hook
