+++ apps/meteor/client/startup/index.ts
@@ -4,7 +4,6 @@ import './appRoot';
 import './audit';
 import './callbacks';
 import './deviceManagement';
-import './e2e';
 import './iframeCommands';
 import './incomingMessages';
 import './messageTypes';
